# Conflicts:
#	src/main/java/com/ferreusveritas/dynamictrees/DynamicTrees.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/IFullGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/IPostGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/IPostGrowFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/IPostRotGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/IPreGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/TreeRegistry.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/configurations/ConfigurationProperty.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/IIfTrueApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/IJsonApplierManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/IPropertyApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/JsonApplierRegistryEvent.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/JsonArrayPropertyApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/JsonPropertyApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/MultiJsonPropertyApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/treepacks/VoidApplier.java
#	src/main/java/com/ferreusveritas/dynamictrees/api/worldgen/BiomePropertySelectors.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/DynamicSaplingBlock.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/FruitBlock.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/NullTreePart.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/branches/TrunkShellBlock.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/leaves/DynamicLeavesBlock.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/leaves/LeavesProperties.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/leaves/LeavesPropertiesManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/rootyblocks/JSONSubAppliersHandler.java
#	src/main/java/com/ferreusveritas/dynamictrees/blocks/rootyblocks/SoilPropertiesManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/CompatHandler.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonGrowthCalculatorActive.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonGrowthCalculatorNull.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonHelper.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonProviderNull.java
#	src/main/java/com/ferreusveritas/dynamictrees/compat/seasons/SeasonProviderSereneSeasons.java
#	src/main/java/com/ferreusveritas/dynamictrees/deserialisation/AxisAlignedBBDeserialiser.java
#	src/main/java/com/ferreusveritas/dynamictrees/deserialisation/BiomeListDeserialiser.java
#	src/main/java/com/ferreusveritas/dynamictrees/deserialisation/JsonPropertyApplierList.java
#	src/main/java/com/ferreusveritas/dynamictrees/deserialisation/ResourceLocationDeserialiser.java
#	src/main/java/com/ferreusveritas/dynamictrees/event/handlers/PoissonDiscEventHandler.java
#	src/main/java/com/ferreusveritas/dynamictrees/growthlogic/JungleLogic.java
#	src/main/java/com/ferreusveritas/dynamictrees/init/DTConfigs.java
#	src/main/java/com/ferreusveritas/dynamictrees/init/DTRegistries.java
#	src/main/java/com/ferreusveritas/dynamictrees/init/DTTrees.java
#	src/main/java/com/ferreusveritas/dynamictrees/resources/DTResourceRegistries.java
#	src/main/java/com/ferreusveritas/dynamictrees/resources/JsonApplierReloadListener.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/dropcreators/drops/Drops.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/dropcreators/drops/StackDrops.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/BeeNestGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/BiomePredicateGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/BottomFlareGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/BushGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/ClearVolumeGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/CocoaGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/ConiferTopperGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/FruitGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/GenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/GenFeatures.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/HugeMushroomGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/HugeMushroomsGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/MoundGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/MushroomRotGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/PodzolGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/RootsGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/RotSoilGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/ShroomlightGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/UndergrowthGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/VinesGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/config/ConfiguredGenFeature.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/nodemappers/DestroyerNode.java
#	src/main/java/com/ferreusveritas/dynamictrees/systems/substances/GrowthSubstance.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/Family.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/FamilyManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/Species.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/SpeciesManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/species/NetherFungusSpecies.java
#	src/main/java/com/ferreusveritas/dynamictrees/trees/species/PalmSpecies.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/AlternateLinkedHashSet.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/CanGrowPredicate.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/ConfiguredGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/FieldGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/ForgeRegistryEntryGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/IJsonObjectGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/JsonHelper.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/JsonObjectGetters.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/ListGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/MapGetter.java
#	src/main/java/com/ferreusveritas/dynamictrees/util/json/ObjectFetchResult.java
#	src/main/java/com/ferreusveritas/dynamictrees/worldgen/BiomeDatabase.java
#	src/main/java/com/ferreusveritas/dynamictrees/worldgen/BiomeDatabaseManager.java
#	src/main/java/com/ferreusveritas/dynamictrees/worldgen/JoCode.java
#	src/main/java/com/ferreusveritas/dynamictrees/worldgen/json/JsonMath.java
#	src/main/java/com/ferreusveritas/dynamictrees/worldgen/json/SpeciesSelectorGetter.java

+++ src/main/java/com/ferreusveritas/dynamictrees/systems/genfeatures/BeeNestGenFeature.java
@@ -1,5 +1,6 @@
 package com.ferreusveritas.dynamictrees.systems.genfeatures;
 
+import com.ferreusveritas.dynamictrees.api.IPostGenFeature;
 import com.ferreusveritas.dynamictrees.api.TreeHelper;
 import com.ferreusveritas.dynamictrees.api.configurations.ConfigurationProperty;
 import com.ferreusveritas.dynamictrees.systems.genfeatures.config.ConfiguredGenFeature;
@@ -47,10 +48,10 @@ public class BeeNestGenFeature extends GenFeature {
     public static final ConfigurationProperty<Block> NEST_BLOCK = ConfigurationProperty.block("nest");
     public static final ConfigurationProperty<WorldGenChanceFunction> WORLD_GEN_CHANCE_FUNCTION = ConfigurationProperty.property("world_gen_chance", WorldGenChanceFunction.class);
 
-    private static final float PLAINS_CHANCE = 0.05f;
-    private static final float FLOWER_FOREST_CHANCE = 0.02f;
-    private static final float FOREST_CHANCE = 0.0005f;
-    private static final float CHANCE = 0.001f;
+    private static final double PLAINS_CHANCE = 0.05f;
+    private static final double FLOWER_FOREST_CHANCE = 0.02f;
+    private static final double FOREST_CHANCE = 0.0005f;
+    private static final double CHANCE = 0.001f;
 
     public BeeNestGenFeature(ResourceLocation registryName) {
         super(registryName);
@@ -86,7 +87,7 @@ public class BeeNestGenFeature extends GenFeature {
             if (BiomeDictionary.hasType(biomeKey, BiomeDictionary.Type.FOREST)) {
                 return FOREST_CHANCE;
             }
-            return 0F;
+            return 0D;
         }).with(MAX_COUNT, 1);
     }
 
@@ -225,7 +226,7 @@ public class BeeNestGenFeature extends GenFeature {
         return validSpaces;
     }
 
-    public interface WorldGenChanceFunction extends BiFunction<IWorld, BlockPos, Float> {
+    public interface WorldGenChanceFunction extends BiFunction<IWorld, BlockPos, Double> {
     }
 
 }
